Separating photo metadata class into own file.
